Decompile func0f0c07c8

+++ src/include/lib/lib_13900.h
@@ -34,7 +34,7 @@ u32 func00014b50(void);
 u32 func00014c98(void);
 s32 func00014e10(u32 arg0);
 s32 func00014eb8(u32 arg0);
-s32 func00014f60(u32 arg0, u16 arg1);
+s32 func00014f60(s8 arg0, u16 arg1);
 s32 func00015020(u32 arg0, u16 arg1);
 u32 func000150c4(void);
 u32 func000150e8(void);
